src/test/javascript/spec/app/entities/ocr-extraction-engine-info/ocr-extraction-engine-info.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { OCRExtractionEngineInfoService } from 'app/entities/ocr-extraction-engine-info/ocr-extraction-engine-info.service';
import { IOCRExtractionEngineInfo, OCRExtractionEngineInfo } from 'app/shared/model/ocr-extraction-engine-info.model';

describe('Service Tests', () => {
  describe('OCRExtractionEngineInfo Service', () => {
    let injector: TestBed;
    let service: OCRExtractionEngineInfoService;
    let httpMock: HttpTestingController;
    let elemDefault: IOCRExtractionEngineInfo;
    let expectedResult: IOCRExtractionEngineInfo | IOCRExtractionEngineInfo[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(OCRExtractionEngineInfoService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new OCRExtractionEngineInfo(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', currentDate, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a OCRExtractionEngineInfo', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            createdAt: currentDate.format(DATE_TIME_FORMAT),
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.create(new OCRExtractionEngineInfo()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a OCRExtractionEngineInfo', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            userName: 'BBBBBB',
            password: 'BBBBBB',
            apiKey: 'BBBBBB',
            createdAt: currentDate.format(DATE_TIME_FORMAT),
            createdBy: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of OCRExtractionEngineInfo', () => {
        const returnedFromService = Object.assign(
          {
            name: 'BBBBBB',
            userName: 'BBBBBB',
            password: 'BBBBBB',
            apiKey: 'BBBBBB',
            createdAt: currentDate.format(DATE_TIME_FORMAT),
            createdBy: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            createdAt: currentDate,
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a OCRExtractionEngineInfo', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
